Move more functionality from both NLU and Embeddings Providers
