+++ includes/Classifai/Features/Classification.php
@@ -14,6 +14,7 @@ use function Classifai\get_post_types_for_language_settings;
 use function Classifai\check_term_permissions;
 use function Classifai\get_classification_feature_enabled;
 use function Classifai\get_classification_feature_taxonomy;
+use function Classifai\get_asset_info;
 
 /**
  * Class Classification
@@ -56,7 +57,12 @@ class Classification extends Feature {
 	 * Set up necessary hooks.
 	 */
 	public function feature_setup() {
-		// var_dump( $this->get_settings() ); die;
+		add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_admin_assets' ] );
+		add_action( 'enqueue_block_editor_assets', [ $this, 'enqueue_editor_assets' ] );
+		add_action( 'classifai_after_feature_settings_form', [ $this, 'render_previewer' ] );
+		add_action( 'rest_api_init', [ $this, 'add_process_content_meta_to_rest_api' ] );
+		add_action( 'add_meta_boxes', [ $this, 'add_meta_box' ], 10, 2 );
+		add_action( 'save_post', [ $this, 'save_metabox' ] );
 	}
 
 	/**
@@ -187,12 +193,253 @@ class Classification extends Feature {
 	}
 
 	/**
+	 * Enqueue the admin scripts.
+	 */
+	public function enqueue_admin_assets() {
+		wp_enqueue_script(
+			'classifai-language-processing-script',
+			CLASSIFAI_PLUGIN_URL . 'dist/language-processing.js',
+			get_asset_info( 'language-processing', 'dependencies' ),
+			get_asset_info( 'language-processing', 'version' ),
+			true
+		);
+
+		wp_enqueue_style(
+			'classifai-language-processing-style',
+			CLASSIFAI_PLUGIN_URL . 'dist/language-processing.css',
+			array(),
+			get_asset_info( 'language-processing', 'version' ),
+			'all'
+		);
+	}
+
+	/**
+	 * Enqueue editor assets.
+	 */
+	public function enqueue_editor_assets() {
+		global $post;
+
+		wp_enqueue_script(
+			'classifai-editor',
+			CLASSIFAI_PLUGIN_URL . 'dist/editor.js',
+			get_asset_info( 'editor', 'dependencies' ),
+			get_asset_info( 'editor', 'version' ),
+			true
+		);
+
+		if ( empty( $post ) ) {
+			return;
+		}
+
+		wp_enqueue_script(
+			'classifai-gutenberg-plugin',
+			CLASSIFAI_PLUGIN_URL . 'dist/gutenberg-plugin.js',
+			array_merge( get_asset_info( 'gutenberg-plugin', 'dependencies' ), array( 'lodash' ) ),
+			get_asset_info( 'gutenberg-plugin', 'version' ),
+			true
+		);
+
+		// TODO: check JS var classifaiEmbeddingData
+
+		wp_add_inline_script(
+			'classifai-gutenberg-plugin',
+			sprintf(
+				'var classifaiPostData = %s;',
+				wp_json_encode(
+					[
+						'NLUEnabled'           => $this->is_feature_enabled(),
+						'supportedPostTypes'   => $this->get_supported_post_types(),
+						'supportedPostStatues' => $this->get_supported_post_statuses(),
+						'noPermissions'        => ! is_user_logged_in() || ! current_user_can( 'edit_post', $post->ID ),
+					]
+				)
+			),
+			'before'
+		);
+	}
+
+	/**
+	 * Add `classifai_process_content` to the REST API for view/edit.
+	 */
+	public function add_process_content_meta_to_rest_api() {
+		$supported_post_types = $this->get_supported_post_types();
+
+		register_rest_field(
+			$supported_post_types,
+			'classifai_process_content',
+			[
+				'get_callback'    => function ( $data ) {
+					$process_content = get_post_meta( $data['id'], '_classifai_process_content', true );
+					return ( 'no' === $process_content ) ? 'no' : 'yes';
+				},
+				'update_callback' => function ( $value, $data ) {
+					$value = ( 'no' === $value ) ? 'no' : 'yes';
+					return update_post_meta( $data->ID, '_classifai_process_content', $value );
+				},
+				'schema'          => [
+					'type'    => 'string',
+					'context' => [ 'view', 'edit' ],
+				],
+			]
+		);
+	}
+
+	/**
+	 * Add metabox to enable/disable language processing.
+	 *
+	 * @param string   $post_type Post type.
+	 * @param \WP_Post $post WP_Post object.
+	 */
+	public function add_meta_box( string $post_type, \WP_Post $post ) {
+		$supported_post_types = $this->get_supported_post_types();
+		$post_statuses        = $this->get_supported_post_statuses();
+		$post_status          = get_post_status( $post );
+
+		if (
+			in_array( $post_type, $supported_post_types, true ) &&
+			in_array( $post_status, $post_statuses, true )
+		) {
+			add_meta_box(
+				'classifai_language_processing_metabox',
+				__( 'ClassifAI Language Processing', 'classifai' ),
+				[ $this, 'render_meta_box' ],
+				null,
+				'side',
+				'high',
+				array( '__back_compat_meta_box' => true )
+			);
+		}
+	}
+
+	/**
+	 * Render metabox content.
+	 *
+	 * @param \WP_Post $post WP_Post object.
+	 */
+	public function render_meta_box( \WP_Post $post ) {
+		wp_nonce_field( 'classifai_language_processing_meta_action', 'classifai_language_processing_meta' );
+		$process_content = get_post_meta( $post->ID, '_classifai_process_content', true );
+		$process_content = ( 'no' === $process_content ) ? 'no' : 'yes';
+		?>
+
+		<p>
+			<label for="classifai-process-content">
+				<input type="checkbox" value="yes" name="_classifai_process_content" id="classifai-process-content" <?php checked( $process_content, 'yes' ); ?> />
+				<?php esc_html_e( 'Automatically tag content on update', 'classifai' ); ?>
+			</label>
+		</p>
+
+		<div class="classifai-clasify-post-wrapper" style="display: none;">
+			<a href="<?php echo esc_url( wp_nonce_url( admin_url( 'admin-post.php?action=classifai_classify_post&post_id=' . $post->ID ), 'classifai_classify_post_action', 'classifai_classify_post_nonce' ) ); ?>" class="button button-classify-post">
+				<?php esc_html_e( 'Suggest terms & tags', 'classifai' ); ?>
+			</a>
+		</div>
+		<?php
+	}
+
+	/**
+	 * Handles saving the metabox.
+	 *
+	 * @param int $post_id Current post ID.
+	 */
+	public function save_metabox( int $post_id ) {
+		if (
+			wp_is_post_autosave( $post_id ) ||
+			wp_is_post_revision( $post_id ) ||
+			! current_user_can( 'edit_post', $post_id )
+		) {
+			return;
+		}
+
+		if (
+			empty( $_POST['classifai_language_processing_meta'] ) ||
+			! wp_verify_nonce( sanitize_text_field( wp_unslash( $_POST['classifai_language_processing_meta'] ) ), 'classifai_language_processing_meta_action' )
+		) {
+			return;
+		}
+
+		$classifai_process_content = isset( $_POST['_classifai_process_content'] ) ? sanitize_key( wp_unslash( $_POST['_classifai_process_content'] ) ) : '';
+
+		if ( 'yes' !== $classifai_process_content ) {
+			update_post_meta( $post_id, '_classifai_process_content', 'no' );
+		} else {
+			update_post_meta( $post_id, '_classifai_process_content', 'yes' );
+		}
+	}
+
+	/**
+	 * Renders the previewer window for the feature.
+	 *
+	 * @param string $active_feature The active feature.
+	 */
+	public function render_previewer( string $active_feature ) {
+		if ( self::ID !== $active_feature || ! $this->is_feature_enabled() ) {
+			return;
+		}
+
+		$provider_instance       = $this->get_feature_provider_instance();
+		$nlu_features            = array();
+		$supported_post_statuses = $this->get_supported_post_statuses();
+		$supported_post_types    = $this->get_supported_post_types();
+
+		$posts_to_preview = get_posts(
+			array(
+				'post_type'      => $supported_post_types,
+				'post_status'    => $supported_post_statuses,
+				'posts_per_page' => 10,
+			)
+		);
+
+		if ( ! empty( $provider_instance->nlu_features ) ) {
+			$nlu_features = $provider_instance->nlu_features;
+		}
+		?>
+
+		<div id="classifai-post-preview-app">
+			<h2><?php esc_html_e( 'Preview Language Processing', 'classifai' ); ?></h2>
+
+			<div id="classifai-post-preview-controls">
+				<select id="classifai-preview-post-selector">
+					<?php foreach ( $posts_to_preview as $post ) : ?>
+						<option value="<?php echo esc_attr( $post->ID ); ?>"><?php echo esc_html( $post->post_title ); ?></option>
+					<?php endforeach; ?>
+				</select>
+
+				<?php wp_nonce_field( 'classifai-previewer-action', 'classifai-previewer-nonce' ); ?>
+
+				<button type="button" class="button" id="get-classifier-preview-data-btn">
+					<span><?php esc_html_e( 'Preview', 'classifai' ); ?></span>
+				</button>
+			</div>
+
+			<div id="classifai-post-preview-wrapper">
+				<?php
+				foreach ( $nlu_features as $feature_slug => $feature ) :
+					if ( ! get_classification_feature_enabled( $feature_slug ) ) {
+						continue;
+					}
+					?>
+
+					<div class="tax-row tax-row--<?php echo esc_attr( $feature_slug ); ?>">
+						<div class="tax-type"><?php echo esc_html( $feature['feature'] ); ?></div>
+					</div>
+
+					<?php
+				endforeach;
+				?>
+			</div>
+		</div>
+
+		<?php
+	}
+
+	/**
 	 * Get the description for the enable field.
 	 *
 	 * @return string
 	 */
 	public function get_enable_description(): string {
-		return esc_html__( 'Enables the automatic content classification of posts.', 'classifai' );
+		return esc_html__( 'Enables automatic content classification.', 'classifai' );
 	}
 
 	/**
